Fold the hinge tests into test_hinge_finder.py, collectivity into its own module
Per review: the hinge regressions belong with the existing hinge tests
rather than in a new file, and collectivity is general enough to want a
test_dynamics_analysis.py of its own for that kind of thing.

No test bodies changed. All four still fail against the pre-fix
analysis.py and pass after it.
