+++ prody/tests/dynamics/test_hinge_finder.py
@@ -1,9 +1,24 @@
 import unittest
+from unittest.mock import patch
+
 import numpy as np
 from prody import *
 
 # TODO: Import your function here
 # from your_module_name import getGlobalHinges
+
+
+def _make_two_chain_atoms(n1, n2):
+    """Tiny synthetic 2-chain AtomGroup (CA-only) for hinge tests."""
+    n = n1 + n2
+    ag = AtomGroup('synthetic')
+    ag.setCoords(np.zeros((n, 3)))
+    ag.setChids(['A'] * n1 + ['B'] * n2)
+    ag.setResnums(list(range(1, n1 + 1)) + list(range(1, n2 + 1)))
+    ag.setNames(['CA'] * n)
+    ag.setResnames(['ALA'] * n)
+    return ag
+
 
 # --- THE UNIT TEST SUITE ---
 class TestHingesWithRealGNM(unittest.TestCase):
@@ -78,6 +93,70 @@ class TestHingesWithRealGNM(unittest.TestCase):
         # Verify we didn't lose Valid hinges from the middle
         # (This assumes the un-trimmed version had middle hinges, which 1AKE Mode 1 usually does)
         self.assertGreater(len(hinges_trimmed), 0, "1AKE Mode 1 should have valid central hinges")
+
+
+class TestGetHingesEmptyRegions(unittest.TestCase):
+    """getHinges must not crash when there is no sign change.
+
+    ``regs`` ends up empty and the old code indexed ``regs[0]``.
+    """
+
+    def test_no_sign_change_returns_empty_list(self):
+        v = np.array([1.0, 1.0, 1.0, 1.0])
+        # Should return [] (no hinges), not raise IndexError.
+        hinges = getHinges(v)
+        self.assertEqual(hinges, [])
+
+    def test_normal_case_still_works(self):
+        # Sanity check the fix doesn't break the normal path.
+        v = np.array([-1, -1, -1, 1, 1, 1.0])
+        self.assertEqual(getHinges(v), [2])
+
+
+class TestGetGlobalHingesChainAccumulation(unittest.TestCase):
+    """Hinges from all chains but the last were silently dropped.
+
+    The accumulation block sat outside the per-chain loop.
+    """
+
+    def test_all_chains_contribute_hinges(self):
+        ag = _make_two_chain_atoms(6, 5)
+
+        # Chain A (global idx 0-5): crossing well inside -> local hinge 2
+        v_a = np.array([-1, -1, -1, 1, 1, 1.0])
+        # Chain B (global idx 6-10): crossing well inside -> local hinge 1
+        v_b = np.array([1, 1, -1, -1, -1.0])
+        vecs = np.concatenate([v_a, v_b])[:, np.newaxis]
+
+        with patch('prody.dynamics.analysis._getModeVecs', return_value=vecs):
+            hinges = getGlobalHinges(object(), atoms=ag, threshold=15,
+                                     space=None, trim=False)
+
+        # Expect a hinge from BOTH chains: global 2 (chain A) and 7 (chain B).
+        self.assertEqual(hinges, [[2, 7]])
+
+
+class TestGetGlobalHingesOffByOne(unittest.TestCase):
+    """``vecs[fst:lst]`` should be ``vecs[fst:lst+1]``.
+
+    ``chains`` stores inclusive (fst, lst) pairs, so the exclusive slice
+    dropped the last atom of every chain from consideration.
+    """
+
+    def test_hinge_at_last_residue_of_chain_is_detected(self):
+        # Single chain so the accumulation-indentation bug can't interfere
+        # (loop runs exactly once) -- isolates the slicing off-by-one.
+        ag = _make_two_chain_atoms(6, 0)  # single 6-atom chain 'A'
+        # crossings at local (1,2) and (4,5); the second one needs the very
+        # last atom of the chain (global index 5) to be included in the slice.
+        v = np.array([-1, -1, 1, 1, 1, -1.0])
+        vecs = v[:, np.newaxis]
+
+        with patch('prody.dynamics.analysis._getModeVecs', return_value=vecs):
+            hinges = getGlobalHinges(object(), atoms=ag, threshold=15,
+                                     space=None, trim=False)
+
+        self.assertEqual(hinges, [[1, 4]])
 
 
 if __name__ == '__main__':
